Always resume engine (#5974)
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2920

When a new dataset is created, the engines are restarted. But as the dataMode is true, the resumeEmgines function does not resume the engine, and especially the _stopProcessing is not reset to false.
As _stopProcessing stays false, no engine is called to process the analysis.
But the resumeEngines is called either with
. processFilterScript
. file is closed (it calls cleanRestart that calls resumeEngines(
. when a new language is chosen.
In this 3 cases , there is no reason why the resumeEngines should be blocked in data mode.