Don't refresh datasetpackage when setting new computed column type
Fixes https://github.com/jasp-stats/jasp-issues/issues/3913

When closing the workspace, the analyses are removed, and then the columns generated by the analyses (but not computed) are set to be just not computed column ( not linked to an analysis). This operation triggers the DataSetPackage to be refreshed, and sometimes this refresh crashes in the middle of Qt calls. But I don't see the need to refresh the DataSetPackage at all, so I just removed it, and since that the crashes does not occur anymore.

I have also replaced the disconect() calls to blockSignals(true) in the QML controls: the disconnect call was there to prevent some signals to be triggered though the JASP control being deleted. This disconnect is not recommended by Qt, and results in a warning that can cause a mutex deadlock in debug mode. So I just replace the disconnect by blockSignals